+++ sql/llx_mahnwesen_profile.key.sql
@@ -1,0 +1,1 @@
+ALTER TABLE llx_mahnwesen_profile ADD UNIQUE INDEX uk_mahnwesen_profile_code (entity, code);